@@ -5,14 +5,17 @@ author:
 homepage:
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
-
+  sdk: '>=3.1.2 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  rxdart: ^0.20.0
+  rxdart: ^0.27.7
 
+
+  # The following adds the Cupertino Icons font to your application.
+  # Use with the CupertinoIcons class for iOS style icons.
+  cupertino_icons: ^1.0.6
 
 dev_dependencies:
   flutter_test:
