@@ -7,8 +7,7 @@ import 'package:rxdart/rxdart.dart';
 
 typedef void OnTapCallback(String value);
 
-class AutoCompleteTextView extends StatefulWidget
-    with AutoCompleteTextInterface {
+class AutoCompleteTextView extends StatefulWidget {
   final double maxHeight;
   final TextEditingController controller;
 
@@ -28,15 +27,15 @@ class AutoCompleteTextView extends StatefulWidget
   final Function focusLost;
   final int suggestionsApiFetchDelay;
   final Function onValueChanged;
-  final Function onSubmitted;
+  final Function(String) onSubmitted;
   final TextInputAction tfInputAction;
   final FocusNode focusNode;
-  
+
   AutoCompleteTextView(
-      {@required this.controller,
-      @required this.focusNode,
+      {required this.controller,
+      required this.focusNode,
       this.onTapCallback,
-      this.onSubmitted,
+      required this.onSubmitted,
       this.maxHeight = 200,
       this.tfCursorColor = Colors.white,
       this.tfCursorWidth = 2.0,
@@ -47,11 +46,11 @@ class AutoCompleteTextView extends StatefulWidget
       this.tfInputAction = TextInputAction.unspecified,
       this.suggestionStyle = const TextStyle(color: Colors.black),
       this.suggestionTextAlign = TextAlign.left,
-      @required this.getSuggestionsMethod,
-      this.focusGained,
+      required this.getSuggestionsMethod,
+      required this.focusGained,
       this.suggestionsApiFetchDelay = 0,
-      this.focusLost,
-      this.onValueChanged});
+      required this.focusLost,
+      required this.onValueChanged});
   @override
   _AutoCompleteTextViewState createState() => _AutoCompleteTextViewState();
 
@@ -64,11 +63,11 @@ class AutoCompleteTextView extends StatefulWidget
 
 class _AutoCompleteTextViewState extends State<AutoCompleteTextView> {
   ScrollController scrollController = ScrollController();
-  OverlayEntry _overlayEntry;
+  OverlayEntry? _overlayEntry;
   LayerLink _layerLink = LayerLink();
   final suggestionsStreamController = new BehaviorSubject<List<String>>();
-  List<String> suggestionShowList = List<String>();
-  Timer _debounce;
+  List<String> suggestionShowList = <String>[];
+  Timer? _debounce;
   bool isSearching = true;
   @override
   void initState() {
@@ -77,12 +76,12 @@ class _AutoCompleteTextViewState extends State<AutoCompleteTextView> {
       if (widget.focusNode.hasFocus) {
         this.isSearching = true;
         this._overlayEntry = this._createOverlayEntry();
-        Overlay.of(context).insert(this._overlayEntry);
-        if(widget?.focusGained != null) widget?.focusGained();
+        Overlay.of(context).insert(this._overlayEntry!);
+        if (widget.focusGained != null) widget.focusGained();
         _onSearchChanged();
       } else {
-        this._overlayEntry.remove();
-        if(widget?.focusLost != null) widget?.focusLost();
+        this._overlayEntry?.remove();
+        if (widget.focusLost != null) widget.focusLost();
       }
     });
     widget.controller.addListener(_onSearchChanged);
@@ -90,10 +89,11 @@ class _AutoCompleteTextViewState extends State<AutoCompleteTextView> {
 
   _onSearchChanged() {
     if (_debounce?.isActive ?? false) {
-      _debounce.cancel();
+      _debounce?.cancel();
     }
-    
-    _debounce = Timer(Duration(milliseconds: widget.suggestionsApiFetchDelay), () {
+
+    _debounce =
+        Timer(Duration(milliseconds: widget.suggestionsApiFetchDelay), () {
       if (isSearching == true) {
         _getSuggestions(widget.controller.text);
       }
@@ -101,64 +101,65 @@ class _AutoCompleteTextViewState extends State<AutoCompleteTextView> {
   }
 
   _getSuggestions(String data) async {
-    List<String> list = await widget.getSuggestionsMethod(data ?? "");
+    List<String> list = await widget.getSuggestionsMethod(data);
     suggestionsStreamController.sink.add(list);
-}
+  }
 
   OverlayEntry _createOverlayEntry() {
-    RenderBox renderBox = context.findRenderObject();
-    var size = renderBox.size;
+    RenderBox? renderBox = context.findRenderObject() as RenderBox?;
+    var size = renderBox?.size;
     return OverlayEntry(
-      builder: (context) => Positioned(
-        width: size.width,
-        child: CompositedTransformFollower(
-          link: this._layerLink,
-          showWhenUnlinked: false,
-          offset: Offset(0.0, size.height + 5.0),
-          child: Material(
-            elevation: 4.0,
-            child: StreamBuilder<Object>(
-              stream: suggestionsStreamController.stream,
-              builder: (context, suggestionData) {
-                if (suggestionData.hasData) {
-                  suggestionShowList = suggestionData.data;
-                  return ConstrainedBox(
-                    constraints: BoxConstraints(
-                      maxHeight: 200,
-                    ),
-                    child: ListView.builder(
-                      controller: scrollController,
-                      padding: EdgeInsets.zero,
-                      shrinkWrap: true,
-                      itemCount: suggestionShowList.length,
-                      itemBuilder: (context, index) {
-                        return ListTile(
-                          title: Text(
-                            suggestionShowList[index],
-                            style: widget.suggestionStyle,
-                            textAlign: widget.suggestionTextAlign,
-                          ),
-                          onTap: () {
-                            isSearching = false;
-                            widget.controller.text = suggestionShowList[index];
-                            suggestionsStreamController.sink.add([]);
-                            if (widget.focusNode.hasPrimaryFocus) {
-                              widget.focusNode.unfocus();
-                            }
-                            widget.onTappedSuggestion(widget.controller.text);
-                          },
-                        );
-                      }
-                    ),
-                  );
-                } else {
-                  return Container();
-                }
-              }
-            ),
-          ),
-        ),
-    ));
+        builder: (context) => Positioned(
+              width: size?.width,
+              child: CompositedTransformFollower(
+                link: this._layerLink,
+                showWhenUnlinked: false,
+                offset: Offset(0.0, size?.height ?? 0 + 5.0),
+                child: Material(
+                  elevation: 4.0,
+                  child: StreamBuilder<Object>(
+                      stream: suggestionsStreamController.stream,
+                      builder: (context, suggestionData) {
+                        if (suggestionData.hasData) {
+                          suggestionShowList =
+                              suggestionData.data as List<String>;
+                          return ConstrainedBox(
+                            constraints: BoxConstraints(
+                              maxHeight: 200,
+                            ),
+                            child: ListView.builder(
+                                controller: scrollController,
+                                padding: EdgeInsets.zero,
+                                shrinkWrap: true,
+                                itemCount: suggestionShowList?.length,
+                                itemBuilder: (context, index) {
+                                  return ListTile(
+                                    title: Text(
+                                      suggestionShowList[index],
+                                      style: widget.suggestionStyle,
+                                      textAlign: widget.suggestionTextAlign,
+                                    ),
+                                    onTap: () {
+                                      isSearching = false;
+                                      widget.controller.text =
+                                          suggestionShowList[index];
+                                      suggestionsStreamController.sink.add([]);
+                                      if (widget.focusNode.hasPrimaryFocus) {
+                                        widget.focusNode.unfocus();
+                                      }
+                                      widget.onTappedSuggestion(
+                                          widget.controller.text);
+                                    },
+                                  );
+                                }),
+                          );
+                        } else {
+                          return Container();
+                        }
+                      }),
+                ),
+              ),
+            ));
   }
 
   @override
@@ -177,7 +178,7 @@ class _AutoCompleteTextViewState extends State<AutoCompleteTextView> {
         onSubmitted: widget.onSubmitted,
         textInputAction: widget.tfInputAction,
         onChanged: (text) {
-          if(widget?.onValueChanged != null) widget?.onValueChanged(text);
+          if (widget?.onValueChanged != null) widget?.onValueChanged(text);
           isSearching = true;
           scrollController.animateTo(
             0.0,
